samples/samples-requirements.txt
python-dotenv
azure-data-tables